+++ packages/stylelint-config-wordpress/__tests__/media-queries.js
@@ -77,7 +77,7 @@ test("There are warnings with invalid media queries CSS", t => {
     t.is(warnings[0].text, "Expected single space after \":\" (media-feature-colon-space-after)", "correct warning text")
     t.is(warnings[1].text, "Expected single space after \":\" (media-feature-colon-space-after)", "correct warning text")
     t.is(warnings[2].text, "Unexpected whitespace before \":\" (media-feature-colon-space-before)", "correct warning text")
-    t.is(warnings[3].text, "Unexpected missing punctuation within non-boolean media feature (media-feature-no-missing-punctuation)", "correct warning text")
+    t.is(warnings[3].text, "Unexpected missing punctuation (media-feature-no-missing-punctuation)", "correct warning text")
     t.is(warnings[4].text, "Expected single space after range operator (media-feature-range-operator-space-after)", "correct warning text")
     t.is(warnings[5].text, "Expected single space after range operator (media-feature-range-operator-space-after)", "correct warning text")
     t.is(warnings[6].text, "Expected single space before range operator (media-feature-range-operator-space-before)", "correct warning text")